requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14
python-dotenv>=1.0
psycopg2-binary>=2.9
gunicorn>=21.2
whitenoise>=6.7
celery[redis]>=5.3
redis>=5.0
prometheus-client>=0.20
httpx>=0.25
python-telegram-bot>=21.0
websockets>=12.0
ccxt>=4.1
pandas>=2.2
numpy>=1.26
tenacity>=8.2
scikit-learn>=1.4
hmmlearn>=0.3
arch>=7.0
